+++ src/main/java/com/koc/user/domain/UserService.java
@@ -1,6 +1,6 @@
 package com.koc.user.domain;
 
-import com.koc.user.application.repository.UserRepository;
+import com.koc.user.adapter.out.persistence.UserRepository;
 import com.koc.user.infrastructure.Entity.UserEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
